Remove underscore from files at qiskit.qasm
