+++ qiskit/circuit/instruction.py
@@ -31,7 +31,7 @@ from copy import deepcopy
 import sympy
 import numpy
 
-from qiskit.qasm._node import _node
+from qiskit.qasm.node import node
 from qiskit.exceptions import QiskitError
 from qiskit.circuit.classicalregister import ClassicalRegister
 
@@ -64,7 +64,7 @@ class Instruction:
             if isinstance(single_param, sympy.Basic):
                 self.params.append(single_param)
             # example: OpenQASM parsed instruction
-            elif isinstance(single_param, _node.Node):
+            elif isinstance(single_param, node.Node):
                 self.params.append(single_param.sym())
             # example: u3(0.1, 0.2, 0.3)
             elif isinstance(single_param, (int, float)):
