--- qiskit/qasm/_node/__init__.py
@@ -1,35 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright 2017, IBM.
-#
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
-
-"""QASM nodes."""
-from ._barrier import Barrier
-from ._binaryop import BinaryOp
-from ._binaryoperator import BinaryOperator
-from ._cnot import Cnot
-from ._creg import Creg
-from ._customunitary import CustomUnitary
-from ._expressionlist import ExpressionList
-from ._external import External
-from ._gate import Gate
-from ._gatebody import GateBody
-from ._id import Id
-from ._idlist import IdList
-from ._if import If
-from ._indexedid import IndexedId
-from ._intnode import Int
-from ._format import Format
-from ._measure import Measure
-from ._opaque import Opaque
-from ._prefix import Prefix
-from ._primarylist import PrimaryList
-from ._program import Program
-from ._qreg import Qreg
-from ._real import Real
-from ._reset import Reset
-from ._unaryoperator import UnaryOperator
-from ._universalunitary import UniversalUnitary
-from ._nodeexception import NodeException
