+++ qiskit/qasm/node/idlist.py
@@ -6,23 +6,23 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 """
-Node for an OPENQASM expression list.
+Node for an OPENQASM idlist.
 """
-from ._node import Node
+from .node import Node
 
 
-class ExpressionList(Node):
-    """Node for an OPENQASM expression list.
+class IdList(Node):
+    """Node for an OPENQASM idlist.
 
-    children are expression nodes.
+    children is a list of id nodes.
     """
 
     def __init__(self, children):
-        """Create the expression list node."""
-        Node.__init__(self, 'expression_list', children, None)
+        """Create the idlist node."""
+        Node.__init__(self, 'id_list', children, None)
 
     def size(self):
-        """Return the number of expressions."""
+        """Return the length of the list."""
         return len(self.children)
 
     def qasm(self, prec=15):
