+++ qiskit/qasm/node/intnode.py
@@ -6,48 +6,46 @@
 # the LICENSE.txt file in the root directory of this source tree.
 
 """
-Node for an OPENQASM real number.
+Node for an OPENQASM integer.
 """
-from sympy import latex, pi
-from sympy.printing.ccode import ccode
-from ._node import Node
+from sympy import N
+
+from .node import Node
 
 
-class Real(Node):
-    """Node for an OPENQASM real number.
+class Int(Node):
+    """Node for an OPENQASM integer.
 
     This node has no children. The data is in the value field.
     """
 
     def __init__(self, id):
-        """Create the real node."""
+        """Create the integer node."""
         # pylint: disable=redefined-builtin
-        Node.__init__(self, "real", None, None)
+        Node.__init__(self, "int", None, None)
         self.value = id
 
     def to_string(self, indent):
         """Print with indent."""
         ind = indent * ' '
-        print(ind, 'real', self.value)
+        print(ind, 'int', self.value)
 
     def qasm(self, prec=15):
         """Return the corresponding OPENQASM string."""
-        if self.value == pi:
-            return "pi"
-
-        return ccode(self.value, precision=prec)
+        # pylint: disable=unused-argument
+        return "%d" % self.value
 
     def latex(self, prec=15, nested_scope=None):
         """Return the corresponding math mode latex string."""
         # pylint: disable=unused-argument
-        return latex(self.value)
+        return "%d" % self.value
 
     def sym(self, nested_scope=None):
         """Return the correspond symbolic number."""
         # pylint: disable=unused-argument
-        return self.value
+        return N(self.value)
 
     def real(self, nested_scope=None):
         """Return the correspond floating point number."""
         # pylint: disable=unused-argument
-        return float(self.value.evalf())
+        return float(self.value)
