+++ qiskit/qasm/node/node.py
