+++ qiskit/qasm/node/nodeexception.py
