+++ qiskit/qasm/node/opaque.py
@@ -8,38 +8,34 @@
 # pylint: disable=invalid-name
 
 """
-Node for an OPENQASM gate definition.
+Node for an OPENQASM opaque gate declaration.
 """
-from ._node import Node
+from .node import Node
 
 
-class Gate(Node):
-    """Node for an OPENQASM gate definition.
+class Opaque(Node):
+    """Node for an OPENQASM opaque gate declaration.
 
     children[0] is an id node.
-    If len(children) is 3, children[1] is an idlist node,
-    and children[2] is a gatebody node.
-    Otherwise, children[1] is an expressionlist node,
-    children[2] is an idlist node, and children[3] is a gatebody node.
+    If len(children) is 3, children[1] is an expressionlist node,
+    and children[2] is an idlist node.
+    Otherwise, children[1] is an idlist node.
     """
 
     def __init__(self, children):
-        """Create the gate node."""
-        Node.__init__(self, 'gate', children, None)
+        """Create the opaque gate node."""
+        Node.__init__(self, 'opaque', children, None)
         self.id = children[0]
         # The next three fields are required by the symbtab
         self.name = self.id.name
         self.line = self.id.line
         self.file = self.id.file
-
         if len(children) == 3:
-            self.arguments = None
-            self.bitlist = children[1]
-            self.body = children[2]
-        else:
             self.arguments = children[1]
             self.bitlist = children[2]
-            self.body = children[3]
+        else:
+            self.arguments = None
+            self.bitlist = children[1]
 
     def n_args(self):
         """Return the number of parameter expressions."""
@@ -53,9 +49,8 @@ class Gate(Node):
 
     def qasm(self, prec=15):
         """Return the corresponding OPENQASM string."""
-        string = "gate " + self.name
+        string = "opaque %s" % self.name
         if self.arguments is not None:
             string += "(" + self.arguments.qasm(prec) + ")"
-        string += " " + self.bitlist.qasm(prec) + "\n"
-        string += "{\n" + self.body.qasm(prec) + "}"
+        string += " " + self.bitlist.qasm(prec) + ";"
         return string
